adapter: move `dataflows` from `coord` to `optimize`
This is the next step of the effort to consolidate all optimizer-related
code in the `mz-adapter` crate into a single module that ultimately be
split into its own crate.
