+++ .github/CODEOWNERS
@@ -23,7 +23,6 @@
 /src/adapter                        @MaterializeInc/adapter
 /src/adapter-types                  @MaterializeInc/adapter
 /src/adapter/src/optimizer          @MaterializeInc/compute
-/src/adapter/src/coord/dataflows.rs @MaterializeInc/compute
 # to track changes to feature flags
 /src/adapter/src/coord/ddl.rs       @MaterializeInc/testing
 # to track changes to feature flags
